src/main/java/com/ssafy/enjoytrip/domain/Trip.java
package com.ssafy.enjoytrip.domain;

import java.util.Date;

public class Trip {
    private Integer id;
    private String name;
    private Date startDate;
    private Date endDate;
    private String tripOverview;
    private String imgUrl;
    private Boolean isPublic;

    // Getters and Setters
    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Date getStartDate() {
        return startDate;
    }

    public void setStartDate(Date startDate) {
        this.startDate = startDate;
    }

    public Date getEndDate() {
        return endDate;
    }

    public void setEndDate(Date endDate) {
        this.endDate = endDate;
    }

    public String getTripOverview() {
        return tripOverview;
    }

    public void setTripOverview(String tripOverview) {
        this.tripOverview = tripOverview;
    }

    public String getImgUrl() {
        return imgUrl;
    }

    public void setImgUrl(String imgUrl) {
        this.imgUrl = imgUrl;
    }

    public Boolean getIsPublic() {
        return isPublic;
    }

    public void setIsPublic(Boolean isPublic) {
        this.isPublic = isPublic;
    }
}